[Feat] - 즐겨찾기 API 4개 추가
