+++ app/src/main/java/com/zion830/threedollars/datasource/UserDataSourceImpl.kt
@@ -8,6 +8,7 @@ import com.zion830.threedollars.datasource.model.v2.response.my.*
 import com.zion830.threedollars.datasource.model.v2.response.visit_history.MyVisitHistoryResponse
 import com.zion830.threedollars.datasource.model.v4.user.request.SignUpRequest
 import com.zion830.threedollars.datasource.model.v4.device.PushInformationRequest
+import com.zion830.threedollars.datasource.model.v4.favorite.request.FavoriteInfoRequest
 import com.zion830.threedollars.datasource.model.v4.medal.request.UpdateMedalRequest
 import com.zion830.threedollars.datasource.model.v4.user.request.EditNameRequest
 import com.zion830.threedollars.network.ServiceApi
