--- app/src/main/java/com/zion830/threedollars/datasource/model/v2/request/FavoriteInfoRequest.kt
@@ -1,10 +1,0 @@
-package com.zion830.threedollars.datasource.model.v2.request
-
-import com.google.gson.annotations.SerializedName
-
-data class FavoriteInfoRequest(
-    @SerializedName("introduction")
-    val introduction: String,
-    @SerializedName("name")
-    val name: String
-)
