+++ app/src/main/java/com/zion830/threedollars/datasource/model/v4/favorite/request/FavoriteInfoRequest.kt
@@ -1,0 +1,10 @@
+package com.zion830.threedollars.datasource.model.v4.favorite.request
+
+import com.google.gson.annotations.SerializedName
+
+data class FavoriteInfoRequest(
+    @SerializedName("introduction")
+    val introduction: String,
+    @SerializedName("name")
+    val name: String
+)
