+++ app/src/main/java/com/zion830/threedollars/ui/favorite/FavoriteMyInfoEditViewModel.kt
@@ -5,7 +5,7 @@ import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.viewModelScope
 import com.zion830.threedollars.R
 import com.zion830.threedollars.datasource.UserDataSource
-import com.zion830.threedollars.datasource.model.v2.request.FavoriteInfoRequest
+import com.zion830.threedollars.datasource.model.v4.favorite.request.FavoriteInfoRequest
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.launch
 import zion830.com.common.base.BaseViewModel
